fix(scroll): fixed a bug with local scrolling
When the page was scrolled all the way to the top, the local
scrolling didn't work for some reason in Chrome. Workaround was,
apparently, to just force the page to scroll by 1px, which made
the issue go away. Seems like a janky fix, but it's working, so
fuck it I guess.